incident-response/requirements.txt
transformers
torch
pinecone-client
langchain
flask
schedule

